scraper/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
python-dotenv==1.1.1
playwright==1.54.0
beautifulsoup4==4.13.4
redis==6.4.0
